+++ src/app/views/PersonFeed.svelte
@@ -5,16 +5,17 @@
   import Content from "src/partials/Content.svelte"
   import Anchor from "src/partials/Anchor.svelte"
   import PersonActions from "src/app/shared/PersonActions.svelte"
-  import {Nip05, Settings, Nip65, Directory} from "src/app/engine"
+  import {Settings, Nip65, Directory} from "src/app/engine"
   import PersonCircle from "src/app/shared/PersonCircle.svelte"
   import PersonAbout from "src/app/shared/PersonAbout.svelte"
   import PersonNotes from "src/app/shared/PersonNotes.svelte"
   import PersonStats from "src/app/shared/PersonStats.svelte"
+  import PersonHandle from "src/app/shared/PersonHandle.svelte"
+  import PersonName from "src/app/shared/PersonName.svelte"
   import {routes} from "src/app/state"
 
   export let pubkey
 
-  const handle = Nip05.handles.key(pubkey)
   const profile = Directory.profiles.key(pubkey).derived(defaultTo({pubkey}))
   const relays = Nip65.getPubkeyHints(Settings.getSetting("relay_limit"), pubkey)
 
@@ -43,19 +44,10 @@
     <div class="flex flex-grow flex-col gap-4">
       <div class="flex items-start justify-between gap-4">
         <div class="flex flex-grow flex-col gap-2">
-          <div class="flex items-center gap-2">
-            <Anchor href={routes.person(pubkey)}>
-              <h1 class="text-2xl">
-                {Directory.displayProfile($profile)}
-              </h1>
-            </Anchor>
-          </div>
-          {#if $handle}
-            <div class="flex gap-1 text-sm">
-              <i class="fa fa-user-check text-accent" />
-              <span class="text-gray-1">{Nip05.displayHandle($handle)}</span>
-            </div>
-          {/if}
+          <Anchor href={routes.person(pubkey)}>
+            <PersonName {pubkey} />
+          </Anchor>
+          <PersonHandle {pubkey} />
         </div>
         <PersonActions {pubkey} />
       </div>